pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "twitter-clone"
version = "0.1.0"
description = "A Twitter clone built with Django"
authors = ["Mitchell Harle <mitch104@protonmail.com>"]
readme = "README.md"
packages = [
    { include = "apps" },
    { include = "config" },
]

[tool.poetry.dependencies]
python = ">=3.10"
Django = "5.2"
psycopg2-binary = "^2.9.10"
django-crispy-forms = "2.0"
crispy-bootstrap5 = "0.7"
Pillow = "11.2.1"
python-dotenv = "1.1.0"
django-htmx = "^1.23.0"
celery = "^5.5.2"
redis = "^5.3.0"
django-celery-results = "^2.5.1"
flower = "^2.0.1"
djangorestframework = "^3.16.0"
drf-spectacular = "^0.28.0"
djangorestframework-api-key = "^3.1.0"

[tool.poetry.group.dev.dependencies]
django-stubs = "^5.2.0"
factory-boy = "3.3.2"
ipdb = "^0.13.11"
mypy = "^1.15.0"
pre-commit = "^4.2.0"
pytest = "^7.1.3"
pytest-django = "^4.5.2"
pytest-docker = "^1.0.1"
pytest-mock = "^3.10.0"
pytest-playwright = "^0.7.0"
ruff = "^0.11.7"

[tool.poetry.scripts]
start = "twitter_clone.manage:main"

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["twitter_clone"]

[tool.mypy]
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = false
disallow_untyped_defs = true
ignore_missing_imports = true
mypy_path = [".", "twitter_clone"]
no_implicit_optional = true
plugins = ["mypy_django_plugin.main"]
python_executable = "python"
python_version = "3.11"
warn_no_return = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = [
    "django.*",
    "django_htmx.*",
    "crispy_forms.*",
    "crispy_bootstrap5.*",
]
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "config.settings"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
django_debug_mode = "keep"
addopts = "--durations=20 --import-mode=importlib"
pythonpath = ["."]
testpaths = ["tests"]
